Fix flush indefinite wait.
If there is network connectivity and a socket is connected, but there is currently no route to the host or the host isn't responding, a client flush call can wait indefinitely (until the server responds again). 

Typically this might trigger a watchdog if the server doesn't respond promptly, so we would like to stop the connection instead.